Refactored a bunch of scripts
