+++ Assets/Scripts/Creature.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 645e00fe1630ffe46802038a04a6f848
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
