+++ Assets/Scripts/Demo/NavMeshDemo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3b8e14749b15c440abcfb2094ea8c83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
